# Replace eqc with stream_data