--- a/test/pbt/encode_decode_type_test.exs
+++ b/test/pbt/encode_decode_type_test.exs
@@ -1,5 +1,5 @@
 defmodule Protobuf.EncodeDecodeTypeTest.PropertyGenerator do
-  alias Protobuf.{Encoder}
+  alias Protobuf.Encoder
 
   require Logger
   import Protobuf.Decoder
@@ -12,18 +12,10 @@ defmodule Protobuf.EncodeDecodeTypeTest.PropertyGenerator do
   defmacro make_property(gen_func, field_type, wire_type) do
     quote do
       property unquote(Atom.to_string(field_type)) <> " roundtrip" do
-        forall n <- unquote(gen_func) do
+        check all n <- unquote(gen_func) do
           iodata = Encoder.encode_type(unquote(field_type), n)
           bin = IO.iodata_to_binary(iodata)
-
-          ensure(
-            n ==
-              decode_type(
-                unquote(wire_type),
-                unquote(field_type),
-                bin
-              )
-          )
+          assert n == decode_type(unquote(wire_type), unquote(field_type), bin)
         end
       end
     end
@@ -35,7 +27,7 @@ defmodule Protobuf.EncodeDecodeTypeTest.PropertyGenerator do
   defmacro make_canonical_property(gen_func, field_type, wire_type) do
     quote do
       property unquote(Atom.to_string(field_type)) <> " canonical roundtrip" do
-        forall n <- unquote(gen_func) do
+        check all n <- unquote(gen_func) do
           encoded_val =
             unquote(field_type)
             |> Encoder.encode_type(n)
@@ -53,14 +45,7 @@ defmodule Protobuf.EncodeDecodeTypeTest.PropertyGenerator do
             |> Encoder.encode_type(canonical_val)
             |> IO.iodata_to_binary()
 
-          ensure(
-            canonical_val ==
-              decode_type(
-                unquote(wire_type),
-                unquote(field_type),
-                bin
-              )
-          )
+          assert canonical_val == decode_type(unquote(wire_type), unquote(field_type), bin)
         end
       end
     end
@@ -69,29 +54,38 @@ end
 
 defmodule Protobuf.EncodeDecodeTypeTest do
   use ExUnit.Case
-  use EQC.ExUnit
+  use ExUnitProperties
 
   import Protobuf.EncodeDecodeTypeTest.PropertyGenerator
 
   defp uint32_gen do
-    let(<<x::unsigned-integer-size(32)>> <- binary(4), do: return(x))
+    map(binary(length: 4), fn <<x::unsigned-integer-size(32)>> -> x end)
   end
 
   defp uint64_gen do
-    let(<<x::unsigned-integer-size(64)>> <- binary(8), do: return(x))
+    map(binary(length: 8), fn <<x::unsigned-integer-size(64)>> -> x end)
   end
 
-  make_property(int(), :int32, 0)
-  make_property(largeint(), :int64, 0)
+  defp large_integer do
+    scale(integer(), &(&1 * 10_000))
+  end
+
+  defp natural_number do
+    map(integer(), &abs/1)
+  end
+
+  make_property(integer(), :int32, 0)
+  make_property(large_integer(), :int64, 0)
   make_property(uint32_gen(), :uint32, 0)
   make_property(uint64_gen(), :uint64, 0)
-  make_property(int(), :sint32, 0)
-  make_property(largeint(), :sint64, 0)
-  make_property(bool(), :bool, 0)
+  make_property(integer(), :sint32, 0)
+  make_property(large_integer(), :sint64, 0)
 
-  make_property(nat(), :fixed64, 1)
-  make_property(largeint(), :sfixed64, 1)
-  make_canonical_property(resize(64, real()), :double, 1)
+  make_property(boolean(), :bool, 0)
 
-  make_canonical_property(resize(32, real()), :float, 5)
+  make_property(natural_number(), :fixed64, 1)
+  make_property(large_integer(), :sfixed64, 1)
+
+  make_canonical_property(float(), :double, 1)
+  make_canonical_property(float(), :float, 5)
 end