--- a/test/pbt/encode_decode_varint_test.exs
+++ b/test/pbt/encode_decode_varint_test.exs
@@ -1,25 +1,27 @@
 defmodule Protobuf.EncodeDecodeVarintTest do
-  use ExUnit.Case
-  use EQC.ExUnit
+  use ExUnit.Case, async: true
+  use ExUnitProperties
 
   alias Protobuf.{Encoder, Decoder}
 
   property "varint roundtrip" do
-    forall n <- largeint() do
+    check all n <- large_integer_gen() do
       iodata = Encoder.encode_varint(n)
       bin = IO.iodata_to_binary(iodata)
       [n] = Decoder.decode_varint(bin, :value)
-      ensure(<<n::signed-64>> == <<n::64>>)
+      assert <<n::signed-64>> == <<n::64>>
     end
-  end
-
-  defp neg_gen do
-    let(x <- largeint(), do: return(-abs(x)))
   end
 
   property "encode_varint for negative integers should always be 10 bytes" do
-    forall n <- neg_gen() do
-      ensure(IO.iodata_length(Encoder.encode_varint(n)) == 10)
+    negative_large_integer_gen = map(large_integer_gen(), &(-abs(&1)))
+
+    check all n <- negative_large_integer_gen do
+      assert IO.iodata_length(Encoder.encode_varint(n)) == 10
     end
+  end
+
+  defp large_integer_gen do
+    scale(integer(), &(&1 * 10_000))
   end
 end