Reasons:

1. stream_data is faster: one of the test files went from 0.6s to 0.2s, the other from 0.5s to 0.2s when switching to stream_data
2. We don't need the additional `mix eqc.install --mini` step in order to run the test suite

Thoughts @tony612?