@whatyouhide 2021-04-16
Actually, let me for once take the liberty to apologize rather than ask for permissions 😄, I'll merge this and move on so that I can also setup GitHub Actions to run CI (#148). @tony612 please let me know if this causes any problems and we can always revert or revert in part!